# Optimize required pod affinity (2)