**What type of PR is this?**

/kind feature

**What this PR does / why we need it**:
This is the second PR in optimizing required pod affinity. This PR converts the data structure used to calculate pod affinity from a map of topology-to-list-of-pods to a map of topology-to-pod-counts. This significantly reduces the overhead of creating the map without compromising the predicate performance.

Basically, for each topology, instead of tracking the exact set of existing pods, we just track their count.

This PR builds on #86030. It offers up to 2.3x improvement over #86030. The two PRs combined offer up to 3.7x improvement.

Before
```
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12           1000  19391988  ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-12               1000  29507122  ns/op
```

After #86030
```
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12           1000  12974929  ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-12               1000  18056693  ns/op
```


After this PR
```
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12           1000  8942505   ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-12               1000  7800239   ns/op
```

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```