@ahg-g 2019-12-09
> Is there another PR coming to cache the affinity terms of the incoming pod?

I guess you are referring to the preemption logic? yes, we can cache that in the metadata, but I thought we just wait for PodInfo.